+++ data.h
@@ -11,14 +11,14 @@ for later use by READ.
 */
 class Data : public Program {
 public:
-	Data(std::vector<double> vals);
+	Data(std::vector<float> vals);
 	
    virtual bool execute(int lineNumber, bool next) const;
 	void list(std::ostream& os) const;	// list this line
 	void preExecute() const;			// run before main program execution
 	
 private:
-	std::vector<double> vals;			// doubles to be stored
+	std::vector<float> vals;			// floats to be stored
 };
 
 #endif
